Route posts tag filters through tagged retrieval
