+++ app/blog/render/retrieve/posts.js
@@ -1,4 +1,6 @@
+const Entry = require("models/entry");
 const { getPage } = require("models/entries");
+const fetchTaggedEntries = require("./helpers/fetchTaggedEntries");
 
 module.exports = function (req, res, callback) {
   const blogID = req?.blog?.id;
@@ -11,14 +13,45 @@ module.exports = function (req, res, callback) {
     pathPrefix: res.locals?.path_prefix ?? req?.template?.locals?.path_prefix,
   };
 
-  req.log("Loading page of entries");
-  getPage(blogID, options, (err, entries, pagination) => {
-    if (err) {
-      return callback(err);
-    }
+  const tags = req?.query?.tag || req?.params?.tag || res?.locals?.tag;
 
-    res.locals.pagination = pagination;
-  
-    callback(null, entries);
-  });
+  if (!tags) {
+    req.log("Loading page of entries");
+    return getPage(blogID, options, (err, entries, pagination) => {
+      if (err) {
+        return callback(err);
+      }
+
+      res.locals.pagination = pagination;
+
+      callback(null, entries);
+    });
+  }
+
+  let page = parseInt(options.pageNumber, 10);
+  if (!page || page < 1) page = 1;
+
+  let limit = parseInt(options.pageSize, 10);
+  if (!Number.isFinite(limit)) limit = undefined;
+  if (!limit || limit < 1 || limit > 500) limit = 100;
+
+  const offset = (page - 1) * limit;
+
+  req.log("Loading tagged page of entries");
+  fetchTaggedEntries(
+    blogID,
+    tags,
+    { limit, offset, pathPrefix: options.pathPrefix },
+    (err, result) => {
+      if (err) {
+        return callback(err);
+      }
+
+      Entry.get(blogID, result.entryIDs || [], (entries) => {
+        entries.sort((a, b) => b.dateStamp - a.dateStamp);
+        res.locals.pagination = result.pagination || {};
+        callback(null, entries);
+      });
+    }
+  );
 };
